Refactored assets directory
