+++ assets/icons/DiagonalArrow.vue
