+++ assets/icons/HardenIcon.vue
