+++ assets/icons/HardenLibIcon.vue
