+++ assets/icons/HeimdallLogo.vue
