+++ assets/icons/NormalizeIcon.vue
