+++ components/home/Capabilities.vue
@@ -71,7 +71,7 @@
   
 <script setup>
 
-import DiagonalArrow from 'assets/DiagonalArrow.vue'
+import DiagonalArrow from 'assets/icons/DiagonalArrow.vue'
 
 const props = defineProps({
   capabilities: {
